# optimized deserializing of `ErrorMessage` and related main process code