When analyzing [ruby-lapack](http://cppcheck1.osuosl.org:8000/ruby-lapack) it will rarely use all the available processes because the files are being analyzed very fast and produce a lot of output which requires parsing thus preventing from new processes being spawned fast enough.

I came across this quite a while ago but didn't have an actual case to test it with.

This improves things but there's more to be optimized. Will go into some less straight forward things in other PRs as well as tickets.