@firewave 2022-12-04
This still needs improved errorhandling.