@firewave 2022-12-08
The utilization of the processes is still non-existent from some point but it's a bit better:

```
Benchmark 1: bin/cppcheck --library=posix --library=gnu  --showtime=top5 --check-library --inconclusive --enable=style,information --inline-suppr --suppress=unmatchedSuppression --template=daca2 --debug-warnings --suppress=autoNoType --suppress=valueFlowBailout --suppress=bailoutUninitVar --suppress=symbolDatabaseWarning --suppress=valueFlowBailoutIncompleteVar -D__GNUC__ --platform=unix64 -j5 /mnt/s/Downloads/ruby-lapack-1.8.2 2> /dev/null
```

before:
```
  Time (mean ± σ):     207.301 s ± 10.444 s    [User: 263.947 s, System: 60.336 s]
  Range (min … max):   190.049 s … 214.663 s    5 runs
```

after:
```
  Time (mean ± σ):     175.161 s ±  3.506 s    [User: 224.577 s, System: 54.583 s]
  Range (min … max):   171.116 s … 178.383 s    5 runs
```

after with `-j1`:
```
  Time (mean ± σ):     110.920 s ± 15.305 s    [User: 99.366 s, System: 6.968 s]
  Range (min … max):   97.794 s … 134.904 s    5 runs
```