@firewave 2022-12-08
With some start-up optimization and some more improvements:
```
  Time (mean ± σ):     164.771 s ±  4.923 s    [User: 209.414 s, System: 50.433 s]
  Range (min … max):   159.761 s … 172.304 s    5 runs
```