@firewave 2022-12-08
"Ir count" using only the first 20 files from the package with the same parameter for the main processs:

Clang 14 `193,310,705` -> `101,474,166`